Refactor `linkify-urls-in-code` to avoid duplicate import of features.tsx
1. options depends on linkify-urls-in-code
2. linkify-urls-in-code was a feature file
3. the feature file depended features.tsx
4. features.tsx also tries to collect features
