+++ source/features/shorten-links.tsx
@@ -1,7 +1,7 @@
 import select from 'select-dom';
 import {applyToLink} from 'shorten-repo-url';
 import features from '../libs/features';
-import {linkifiedURLClass} from './linkify-urls-in-code';
+import {linkifiedURLClass} from '../libs/linkify-text-nodes';
 
 function init(): void {
 	for (const a of select.all<HTMLAnchorElement>(`a[href]:not(.${linkifiedURLClass})`)) {
